requirements.txt
streamlit
openai
python-dotenv
fastapi
os